Moved new postgres implementation to v2 folder, also added nil check for db object in postgres config
